Edit html, CSS, and Javascript to add sub button and fix bug
Missing subtraction button was added and removed the zero button's
selector to fit the button into the container while retaining the size.
The bug involving the double-clicking of math operators was solved by
removing the "numberChoice.pop()" function from the conditional
solution. The logic was that the conditional would look at the previous
selection in the array and see if it was a math operator. This would
only work once as the pop function removed the only sign of showing
a previous input based on the conditionals relying on the numberChoice
array.

Ex.
numberChoice: [1], [2], [+]
display:12+
'hit + again'
(activate display with pop)
numberChoice: [1], [2]
'hit + again'
(activate display with pop again)
display: 12++

Because the second conditional check saw that the last number in the
numberChoice array was not a math operator, it allowed the display to
add the extra "+".