Change operator button logic and start logic for operate function
Was still running into issues with more than one operator appearing and
opted to just disable the buttons for the operators with a single click
of any operator. Also adjusted numberChoice array to concatenate it's
containers to reflect the display. Operate function will have the same
logic before applying arithmetic to the values.