Change operate function click to inline HTML and adjust math functions
Was experiencing difficulty with the addEventListener for the button
created with the querySelector. For some reason the operate function was
instantly called without input. Will research on reason later. For now
all buttons are called with the inline html onclick tag. Math functions
are functional for Int value numbers and dividing by zero will show the
error message.