Provide full functionality of the calculator
The non-number buttons are disabled initially to avoid bugs like
entering the operators first or entering a decimal first. All math
functions are fully operable via the buttons assigned to them.
Pressing an operator after the results of a calculation will take
the result and use it as the first value of the mathematics operator.
Console logs the entire step of the mathematic process by showing the
numbers entered into the array and what operator function is used.